CLIP_example/app/requirements.txt
Flask==2.0.2
weaviate_client==3.3.3
Werkzeug==2.0.1
imageio==2.33.1
sage_data_client==0.7.1
matplotlib==3.8.2
Pillow